spread loader and auth reducers
