--- src/Components/Game/types.ts
@@ -1,1 +1,0 @@
-export const TEST_ACTION = "TEST_ACTION";
