+++ src/Components/Loader/types.ts
@@ -1,0 +1,2 @@
+export const LOADING_START = "LOADING_START";
+export const LOADING_END = "LOADING_END";
