+++ src/Components/NameForm/types.ts
@@ -1,5 +1,3 @@
-export const LOADING_START = "LOADING_START";
-export const LOADING_END = "LOADING_END";
 export const LOGIN = "LOGIN";
 export const LOGIN_FAILURE = "LOGIN_FAILURE";
 export const LOGOUT = "LOGOUT";
