+++ src/Routing/Routes.test.tsx
@@ -11,7 +11,6 @@ import { LoginPage } from "../Pages/LoginPage";
 import { GamePage } from "../Pages/GamePage";
 
 import { defaultState as loginDefaultState } from "../Components/NameForm/reducer";
-import { defaultState as gameDefaultState } from "../Components/Game/reducer";
 
 const mockStore = configureMockStore([thunk]);
 const mockHistory = { push: jest.fn() };
